+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### v0.3.2 (dev)
 - Set up generic compiler to generalize classes
+- Simplified step evaluation process
+- Added functions:
+  - Numeric: `Factor`
 ### v0.3.1 (Nov 22, 2021)
 - Set up simple user interface in web
 - Added functions: